% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step4_junction_cluster.R
\name{asp_selection}
\alias{asp_selection}
\title{AEN filter}
\usage{
asp_selection(corm, sample_cutoff = 2, gene_cutoff = 1500, link_cutoff = 10)
}
\arguments{
\item{corm}{Data frame containing ASP-gene relationships with columns: junction, correlation, Freq, etc.}

\item{sample_cutoff}{Minimum number of samples supporting each link (default = 2)}

\item{gene_cutoff}{Number of top junctions to keep based on number of links (default = 1500)}

\item{link_cutoff}{Minimum number of biased (directionally consistent) links per junction (default = 10)}
}
\value{
Filtered high-confidence ASP-gene correlation network (AEN)
}
\description{
Filters non-key links in the AEN network based on frequency and directional bias.
}
