% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step4_junction_cluster.R
\name{gene_clustering}
\alias{gene_clustering}
\title{AEN Clustering}
\usage{
gene_clustering(corm2, cluster_num = 25, asp_num = 10)
}
\arguments{
\item{corm2}{The data frame contating the high quality AEN network with junction as symbol}

\item{cluster_num}{The number of clusters}

\item{asp_num}{The least number of ASPs within ASP Clusters}
}
\value{
a list containing a data frame of ASP clusters and a similarity matrix
}
\description{
The function to detect asp clusters based on the AEN network
}
