% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_merge.R
\name{merge_cor}
\alias{merge_cor}
\title{Merge Correlation}
\usage{
merge_cor(cor_list)
}
\arguments{
\item{cor_list}{the list containing data.frame from the above steps for each sample}
}
\value{
a data frame containing the correlation link (AEN linke) from multiple samples.
}
\description{
The function to merge the correlation across samples from the step1
}
\examples{
merge_cor <- merge_cor(cor_list)
}
