% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1_cor.R
\name{step1_cor}
\alias{step1_cor}
\title{Correlation}
\usage{
step1_cor(
  exp,
  mat,
  annj,
  met,
  pa = "ALL",
  coverage_cutoff = 5,
  cell_cutoff = 20,
  gene_cutoff = 5
)
}
\arguments{
\item{exp}{Cell-Gene Expression Matrix with Cell as columns and Gene as rows}

\item{mat}{Cell-junction Count Matrix with Cell as columns and junction as rows}

\item{annj}{The ASP annotation files from Step 0}

\item{met}{The Cell annotation files with the name of columns : cell and Patient}

\item{pa}{The Name of Patient or Sample chose to detect the correlation between ASP and gene}

\item{coverage_cutoff}{The cutoff of coverage for ASPs}

\item{gene_cutoff}{The number cutoff for genes with the expression within cells > gene_cutoff}

\item{dcell_cutoff}{The number cutoff of cells with coverage > coverage_cutoff}
}
\value{
a data frame containing the information of the relationship between ASP and gene with columns: ASP, Gene, correlation, P value, and Others.
}
\description{
The function to detect the correlation between ASP (Alternative Splicing Pattern) and Gene Expression Network (AEN) within one sample or multiple samples
}
\examples{
cor <- step1_cor(exp,mat,annj,met,pa = "ALL",coverage_cutoff = 5,cell_cutoff = 20,gene_cutoff = 5)
}
