% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step0_asp.R
\name{asp}
\alias{asp}
\title{ASP}
\usage{
asp(desj)
}
\arguments{
\item{desj}{a data frame containing the junction used in the samples with junction as the first column}
}
\value{
a data frame containing the information of ASP as well as the composition of ASP
}
\description{
The function to detect the ASP (Alternative Splicing Pattern)
}
\examples{
asp <- asp(desj)
}
