% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step4_cell_clustering.R
\name{asp_score}
\alias{asp_score}
\title{ASP Clusters enrichment Score}
\usage{
asp_score(asp, ann_junc)
}
\arguments{
\item{asp}{Cell-junction Count Matrix with Cell as columns and junction as rows}

\item{ann_junc}{The ASP Clusters}
}
\value{
Cell-ASP Clusters enrichment Score matrix
}
\description{
The enrichment score of ASP Clusters for each cell
}
\examples{
asp_mat <- asp_score(asp,ann_junc)
}
