% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3_junction_clustering.R
\name{asp_selection}
\alias{asp_selection}
\title{AEN filter}
\usage{
asp_selection(
  corm,
  sample_filter = T,
  sample_cutoff = 2,
  gene_cutoff = 1500,
  link_cutoff = 10
)
}
\arguments{
\item{corm}{The data frame contating the relationship (AEN) between ASP and Gene from the above step 2 or 1.}

\item{sample_cutoff}{The number cutoff of samples for the links within the AEN}

\item{gene_cutoff}{The top key ASPs used to the downstream analysis}

\item{link_cutoff}{the link cutoff for the ASPs}
}
\value{
The high quality AEN
}
\description{
The function to filter the non-key links with AEN network
}
\examples{
AEN_h <- asp_selction(corm,sample_filter = T,sample_cutoff = 2,gene_cutoff = 1500,link_cutoff = 10)
}
