% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step4_cell_clustering.R
\name{cell_clus}
\alias{cell_clus}
\title{Cell Clusters}
\usage{
cell_clus(asp_score, resolution = 0.5, min.dist = 1)
}
\arguments{
\item{asp_score}{Cell-ASP Clusters enrichment Score matrix}

\item{resolution}{The resolution for the function FindClusters}

\item{min.dist}{The min distance for the function RunUMAP}
}
\value{
Cell clusters and the umap determined by ASP
}
\description{
The Cell Clusters determined by ASP Clusters
}
\examples{
cell_info <- cell_clus(asp_score,resolution = 0.5,min.dist = 1)
}
