% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3_junction_clustering.R
\name{junction_clustering}
\alias{junction_clustering}
\title{AEN Clustering}
\usage{
junction_clustering(corm2, cluster_num = 25, asp_num = 10)
}
\arguments{
\item{corm2}{The data frame contating the high quality AEN network}

\item{cluster_num}{The number of clusters}

\item{asp_num}{The least number of ASPs within ASP Clusters}
}
\value{
a data frame containing the information of ASP clusters
}
\description{
The function to detect asp clusters based on the AEN network
}
\examples{
junction_clusters <- junction_clustering(corm2,cluster_num = 25,asp_num = 10)
}
